Rename todo drawer -> task drawer
