+++ frontend/src/app/store.ts
@@ -1,14 +1,14 @@
 import {configureStore, ThunkAction, Action} from '@reduxjs/toolkit';
 import {apiSlice} from "../features/api/apiSlice";
 import authSlice from "../features/auth/authSlice";
-import todoDrawerSlice from "../features/tasks/todoSlice";
+import taskDrawerSlice from "../features/tasks/taskDrawerSlice";
 import scaffoldSlice from "../components/scaffold/scaffoldSlice";
 
 export const store = configureStore({
     reducer: {
         [apiSlice.reducerPath]: apiSlice.reducer,
         auth: authSlice,
-        todoDrawer: todoDrawerSlice,
+        taskDrawer: taskDrawerSlice,
         scaffold: scaffoldSlice,
     },
     middleware: getDefaultMiddleware => getDefaultMiddleware().concat(apiSlice.middleware)
